+++ packages/xrpl/src/models/transactions/vaultDeposit.ts
@@ -1,0 +1,41 @@
+import { Amount } from '../common'
+
+import {
+  BaseTransaction,
+  validateBaseTransaction,
+  validateRequiredField,
+  isString,
+  isAmount,
+} from './common'
+
+/**
+ * The VaultDeposit transaction adds liqudity in exchange for vault shares.
+ *
+ * @category Transaction Models
+ */
+export interface VaultDeposit extends BaseTransaction {
+  TransactionType: 'VaultDeposit'
+
+  /**
+   * The ID of the vault to which the assets are deposited.
+   */
+  VaultID: string
+
+  /**
+   * Asset amount to deposit.
+   */
+  Amount: Amount
+}
+
+/**
+ * Verify the form and type of a {@link VaultDeposit} at runtime.
+ *
+ * @param tx - A {@link VaultDeposit} Transaction.
+ * @throws When the {@link VaultDeposit} is malformed.
+ */
+export function validateVaultDeposit(tx: Record<string, unknown>): void {
+  validateBaseTransaction(tx)
+
+  validateRequiredField(tx, 'VaultID', isString)
+  validateRequiredField(tx, 'Amount', isAmount)
+}