+++ packages/xrpl/test/models/clawback.test.ts
@@ -45,7 +45,7 @@ describe('Clawback', function () {
     assert.throws(
       () => validate(invalidAmount),
       ValidationError,
-      'Clawback: invalid Amount',
+      'Clawback: invalid field Amount',
     )
 
     const invalidStrAmount = {
@@ -57,7 +57,7 @@ describe('Clawback', function () {
     assert.throws(
       () => validate(invalidStrAmount),
       ValidationError,
-      'Clawback: invalid Amount',
+      'Clawback: invalid field Amount',
     )
   })