+++ packages/ripple-binary-codec/HISTORY.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+* Support for `Single Asset Vault` (XLS-65)
+* Adds new `STNumber` serialization type.
+
 ## 2.4.1 (2025-6-18)
 
 ### Fixed